Combined the Document/Media "Workspace Has Collection" condition
