--- src/packages/documents/documents/conditions/index.ts
@@ -1,1 +1,0 @@
-export { UMB_DOCUMENT_WORKSPACE_HAS_COLLECTION_CONDITION } from './document-workspace-has-collection.condition.js';
