--- src/packages/documents/documents/conditions/manifests.ts
@@ -1,3 +1,0 @@
-import { manifest as docummentWorkspaceHasCollectionCondition } from './document-workspace-has-collection.condition.js';
-
-export const manifests = [docummentWorkspaceHasCollectionCondition];
