+++ src/packages/documents/documents/index.ts
@@ -8,7 +8,6 @@ export * from './user-permissions/index.js';
 export * from './components/index.js';
 export * from './entity.js';
 export * from './entity-actions/index.js';
-export * from './conditions/index.js';
 export * from './modals/index.js';
 
 export * from './tree/index.js';
