+++ src/packages/documents/documents/manifests.ts
@@ -1,6 +1,5 @@
 import { manifests as breadcrumbManifests } from './breadcrumb/manifests.js';
 import { manifests as collectionManifests } from './collection/manifests.js';
-import { manifests as conditionManifests } from './conditions/manifests.js';
 import { manifests as entityActionManifests } from './entity-actions/manifests.js';
 import { manifests as entityBulkActionManifests } from './entity-bulk-actions/manifests.js';
 import { manifests as menuItemManifests } from './menu-item/manifests.js';
@@ -16,7 +15,6 @@ import { manifests as workspaceManifests } from './workspace/manifests.js';
 export const manifests = [
 	...breadcrumbManifests,
 	...collectionManifests,
-	...conditionManifests,
 	...entityActionManifests,
 	...entityBulkActionManifests,
 	...menuItemManifests,
