--- src/packages/media/media/conditions/index.ts
@@ -1,1 +1,0 @@
-export { UMB_MEDIA_WORKSPACE_HAS_COLLECTION_CONDITION } from './media-workspace-has-collection.condition.js';
