--- src/packages/media/media/conditions/manifests.ts
@@ -1,3 +1,0 @@
-import { manifest as mediaWorkspaceHasCollectionCondition } from './media-workspace-has-collection.condition.js';
-
-export const manifests = [mediaWorkspaceHasCollectionCondition];
