+++ src/packages/media/media/index.ts
@@ -6,7 +6,6 @@ export * from './tracked-reference/index.js';
 export * from './components/index.js';
 export * from './entity.js';
 export * from './utils/index.js';
-export * from './conditions/index.js';
 
 export { UMB_MEDIA_TREE_ALIAS } from './tree/index.js';
 export { UMB_MEDIA_COLLECTION_ALIAS } from './collection/index.js';
