+++ src/packages/media/media/manifests.ts
@@ -1,5 +1,4 @@
 import { manifests as collectionManifests } from './collection/manifests.js';
-import { manifests as conditionManifests } from './conditions/manifests.js';
 import { manifests as entityActionsManifests } from './entity-actions/manifests.js';
 import { manifests as entityBulkActionsManifests } from './entity-bulk-actions/manifests.js';
 import { manifests as menuItemManifests } from './menu-item/manifests.js';
@@ -11,7 +10,6 @@ import { manifests as workspaceManifests } from './workspace/manifests.js';
 
 export const manifests = [
 	...collectionManifests,
-	...conditionManifests,
 	...entityActionsManifests,
 	...entityBulkActionsManifests,
 	...menuItemManifests,
